Moved lrn_python and l2_normalize_python to topi.testing
