+++ topi/tests/python_cpp/test_topi_l2norm.py
@@ -4,30 +4,7 @@ import tvm
 import topi
 import logging
 from topi.util import get_const_tuple
-
-def l2_normalize_python(a_np, eps, axis=None):
-    """L2 normalize operator in NCHW layout.
-
-    Parameters
-    ----------
-    a_np : numpy.ndarray
-        4-D with shape [batch, in_channel, in_height, in_width]
-
-    eps : float
-        epsilon constant value
-    axis : list of int
-        axis over the normalization applied
-
-    Returns
-    -------
-    l2_normalize_out : np.ndarray
-        4-D with shape [batch, out_channel, out_height, out_width]
-    """
-    dot_value = np.power(a_np, 2.0)
-    sqr_sum = np.sum(dot_value, axis, keepdims=True)
-    sqrt_sum = np.sqrt(np.maximum(np.broadcast_to(sqr_sum, a_np.shape), eps))
-    l2_normalize_out = np.divide(a_np, sqrt_sum)
-    return l2_normalize_out
+import topi.testing
 
 def verify_l2_normalize(shape, eps, axis=None):
     '''Verify l2 normalization operator by comparing outputs from tvm and numpy implementation'''
@@ -36,7 +13,7 @@ def verify_l2_normalize(shape, eps, axis=None):
     dtype = A.dtype
 
     a_np = np.random.uniform(size=shape).astype(dtype)
-    b_np = l2_normalize_python(a_np, eps, axis)
+    b_np = topi.testing.l2_normalize_python(a_np, eps, axis)
 
     def check_device(device):
         if not tvm.module.enabled(device):
